requirements.txt
chainlit==2.4.400
llama_index==0.12.27
pinecone==6.0.2
python-dotenv==1.1.0
semantic_router==0.1.6
llama-index-tools-wikipedia